Manage /etc/aliases with the template module
Using the `lineinfile` module causes the deletion of aliases in the vars
files not to be reflected in the actual `/etc/aliases` file, so we may
end up with dangling entries.

Using a template to manage the whole file makes sense since this file
should be "owned" by the postfix installation/configuration, so we want
it to be rebuilt as a whole from the vars configuration instead of just
editing some lines.

Closes #78

Signed-off-by: L. Alberto Giménez <agimenez@sysvalve.es>